feat(handdecomp): port StartSallyCirProc (sub_8099D84) → matching C
screen NEAR 0e->16 = JP signed-param sign-ext (lsr->asr); int param +
int uu=(s8)unk early-cast forces asr (proto u8->int). matching-C 7640→7641.

+++ include/prepscreen.h
@@ -631,7 +631,7 @@ struct SallyCirProc {
 void SallyCir_Init(struct SallyCirProc * proc);
 void SallyCir_Loop(struct SallyCirProc * proc);
 void SallyCir_OnEnd(void);
-struct SallyCirProc* StartSallyCirProc(ProcPtr parent, u8 unk);
+struct SallyCirProc* StartSallyCirProc(ProcPtr parent, int unk);
 void SallyCir_RectInit(struct SallyCirProc * proc);
 void SallyCir_RectLoop(struct SallyCirProc * proc);
 int PrepUtils_CondFalse_0(void);
